lsif-clang: Reduce contention by using per-table mutexes.
For an indexing job of CMake on a 60 core machine:
- Before this change: indexing time 30s with about 31/60 core utilization
- After this change:  indexing time 25s with about 42/60 core utilization

There doesn't seem to be much effect on 16 core machines.

This was changed in upstream LLVM too (in clangd).
The patch to IndexerMain.cpp is not used, because the code is
duplicated in LSIFClangMain.cpp, but we keep the two in sync.